# packages/server/src/swagger/swagger.yaml
basePath: null
definitions:
...
    KeyValueDto:
        description: ""
        properties:
            label:
                type: string
                description: ""
            value:
                type: string
                description: ""
        type: object
        required:
            - label
            - value
    Query:
        description: ""
        properties:
            size:
                type: number
                format: double
                description: ""
            page:
                type: number
                format: double
                description: ""
        type: object
        required:
            - size
            - page
...
    CreateNoticeReq:
        description: ""
        properties:
            title:
                type: string
                description: ""
            description:
                type: string
                description: ""
            extra:
                type: string
                description: ""
            status:
                type: string
                description: ""
            type:
                type: string
                description: ""
        type: object
        required:
            - title
            - description
            - extra
            - status
            - type
    Notice:
        description: ""
        properties:
            id:
                type: string
                description: ""
            title:
                type: string
                description: ""
            description:
                type: string
                format: date-time
                description: ""
            type:
                type: string
                description: ""
            extra:
                type: string
                description: ""
            read:
                type: boolean
                description: ""
        type: object
        required:
            - id
            - title
            - description
            - type
            - extra
            - read
    EditNoticeReq:
        description: ""
        properties:
            id:
                type: string
                description: ""
            title:
                type: string
                description: ""
            description:
                type: string
                description: ""
            extra:
                type: string
                description: ""
            status:
                type: string
                description: ""
            type:
                type: string
                description: ""
        type: object
        required:
            - id
            - title
            - description
            - extra
            - status
            - type
    ResultList$Notice:
        description: ""
        properties:
            list:
                type: array
                items:
                    $ref: '#/definitions/Notice'
                description: ""
            count:
                type: number
                format: double
                description: ""
            query:
                $ref: '#/definitions/Query'
        type: object
        required:
            - list
    CreateSettingReq:
        description: ""
        properties:
            name:
                type: string
                description: ""
            key:
                type: string
                description: ""
            value:
                type: object
                description: ""
            description:
                type: string
                description: ""
        type: object
        required:
            - name
            - key
            - value
            - description
    Setting:
        description: ""
        properties:
            id:
                type: string
                description: ""
            name:
                type: string
                description: ""
            key:
                type: string
                description: ""
            value:
                type: object
                description: ""
            description:
                type: string
                description: ""
        type: object
        required:
            - id
            - name
            - key
            - value
            - description
    EditSettingReq:
        description: ""
        properties:
            id:
                type: string
                description: ""
            name:
                type: string
                description: ""
            key:
                type: string
                description: ""
            value:
                type: object
                description: ""
            description:
                type: string
                description: ""
        type: object
        required:
            - id
            - name
            - key
            - value
            - description
    SettingsGroup:
        description: ""
        properties: {}
        type: object
    ResultList$Setting:
        description: ""
        properties:
            list:
                type: array
                items:
                    $ref: '#/definitions/Setting'
                description: ""
            count:
                type: number
                format: double
                description: ""
            query:
                $ref: '#/definitions/Query'
        type: object
        required:
            - list
    SettingRes:
        description: ""
        properties:
            id:
                type: string
                description: ""
            name:
                type: string
                description: ""
            key:
                type: string
                description: ""
            value:
                type: object
                description: ""
            description:
                type: string
                description: ""
        type: object
        required:
            - id
            - name
            - key
            - value
            - description
...
info:
    description: 项目标准接口
    license:
        name: MIT
    title: 'rest api interface'
    version: 0.0.1
paths:
...
    /notice/:
        post:
            operationId: NoticesCreate
            produces:
                - text/html
            responses:
                '204':
                    description: 'No content'
            description: ""
            tags:
                - core
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: body
                    name: entry
                    required: true
                    schema:
                        $ref: '#/definitions/CreateNoticeReq'
        put:
            operationId: NoticesUpdate
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/Notice'
            description: ""
            tags:
                - core
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: body
                    name: entry
                    required: true
                    schema:
                        $ref: '#/definitions/EditNoticeReq'
    /notice/search:
        get:
            operationId: NoticesSearch
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/KeyValueDto'
            description: ""
            tags:
                - core
            parameters:
                -
                    description: ""
                    in: query
                    name: keyword
                    required: false
                    type: string
                -
                    description: ""
                    in: query
                    name: value
                    required: false
                    type: string
    /notice/query:
        get:
            operationId: NoticesQuery
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/ResultList$Notice'
            description: ""
            tags:
                - core
            parameters:
                -
                    description: ""
                    in: query
                    name: keyword
                    required: false
                    type: string
                -
                    description: ""
                    in: query
                    name: page
                    required: false
                    type: number
                    format: double
                    default: 1
                -
                    description: ""
                    in: query
                    name: size
                    required: false
                    type: number
                    format: double
                    default: 10
    '/notice/{id}':
        get:
            operationId: NoticesFindOne
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/Notice'
            description: ""
            tags:
                - core
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: string
...
    /setting/:
        post:
            operationId: SettingsCreate
            produces:
                - text/html
            responses:
                '204':
                    description: 'No content'
            description: ""
            tags:
                - core
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: body
                    name: entry
                    required: true
                    schema:
                        $ref: '#/definitions/CreateSettingReq'
        put:
            operationId: SettingsUpdate
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/Setting'
            description: ""
            tags:
                - core
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: body
                    name: entry
                    required: true
                    schema:
                        $ref: '#/definitions/EditSettingReq'
    '/setting/name/{name}':
        put:
            operationId: SettingsUpdateSettingsByName
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/SettingsGroup'
            description: ""
            tags:
                - core
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: query
                    name: name
                    required: true
                    type: string
                -
                    description: ""
                    in: body
                    name: entry
                    required: true
                    schema:
                        $ref: '#/definitions/SettingsGroup'
        get:
            operationId: SettingsGetSettingsByName
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/SettingsGroup'
            description: ""
            tags:
                - core
            parameters:
                -
                    description: ""
                    in: query
                    name: name
                    required: false
                    type: string
    /setting/search:
        get:
            operationId: SettingsSearch
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/KeyValueDto'
            description: ""
            tags:
                - core
            parameters:
                -
                    description: ""
                    in: query
                    name: keyword
                    required: false
                    type: string
                -
                    description: ""
                    in: query
                    name: value
                    required: false
                    type: string
    /setting/query:
        get:
            operationId: SettingsQuery
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/ResultList$Setting'
            description: ""
            tags:
                - core
            parameters:
                -
                    description: ""
                    in: query
                    name: keyword
                    required: false
                    type: string
                -
                    description: ""
                    in: query
                    name: page
                    required: false
                    type: number
                    format: double
                    default: 1
                -
                    description: ""
                    in: query
                    name: size
                    required: false
                    type: number
                    format: double
                    default: 10
    '/setting/key/{key}':
        get:
            operationId: SettingsGetSettingsByKey
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/SettingRes'
            description: ""
            tags:
                - core
            parameters:
                -
                    description: ""
                    in: query
                    name: key
                    required: true
                    type: string
    '/setting/{id}':
        get:
            operationId: SettingsFindOne
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/Setting'
            description: ""
            tags:
                - core
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: string
...
swagger: '2.0'
securityDefinitions: {}
produces:
    - application/json